requirements.txt
openai-whisper 
tqdm
yt-dlp
